[WIP] Separate database methods from keystore.go to PersistentKeystore.go, some refactoring
