+++ crypto/merklekeystore/keystore.go
@@ -17,14 +17,12 @@
 package merklekeystore
 
 import (
-	"encoding/binary"
-	"context"
-	"database/sql"
 	"errors"
 	"fmt"
+
 	"github.com/algorand/go-algorand/crypto"
 	"github.com/algorand/go-algorand/crypto/merklearray"
-	"github.com/algorand/go-algorand/util/db"
+	"github.com/algorand/go-algorand/protocol"
 )
 
 type (
@@ -129,63 +127,6 @@ func New(firstValid, lastValid, interval uint64, sigAlgoType crypto.AlgorithmTyp
 	return s, nil
 }
 
-// Persist dumps the keys into the database and deletes the reference to them in Signer
-func (s *Signer) Persist(store db.Accessor) error {
-	if s.signatureAlgorithms == nil {
-		return fmt.Errorf("no keys provided (nil)")
-	}
-
-	err := store.Atomic(func(ctx context.Context, tx *sql.Tx) error {
-		err := keystoreInstallDatabase(tx) // assumes schema table already exists (created by partInstallDatabase)
-		if err != nil {
-			return err
-		}
-
-		if s.Interval == 0 {
-			return errIntervalZero
-		}
-		round := indexToRound(s.FirstValid, s.Interval, 0)
-		for i, key := range s.signatureAlgorithms {
-			encodedKey := key.MarshalMsg(protocol.GetEncodingBuf())
-			_, err := tx.Exec("INSERT INTO StateProofKeys (id, round, key) VALUES (?,?,?)", i, round, encodedKey)
-			protocol.PutEncodingBuf(encodedKey)
-			if err != nil {
-				return fmt.Errorf("failed to insert StateProof key number %v round %d. SQL Error: %w", i, round, err)
-			}
-			round += s.Interval
-		}
-
-		return nil
-	})
-	if err != nil {
-		return fmt.Errorf("PersistentKeystore.Persist: %w", err)
-	}
-
-	return nil // Success
-}
-
-const keystoreSchemaVersion = 1
-const keystoreTableSchemaName = "merklekeystore"
-
-func keystoreInstallDatabase(tx *sql.Tx) error {
-	_, err := tx.Exec(`CREATE TABLE StateProofKeys (
-    	id	  INTEGER PRIMARY KEY, 
-    	round INTEGER,	    --*  committed round for this key
-		key   BLOB  --*  msgpack encoding of ParticipationAccount.StateProof.GenericSigningKey
-		);`)
-	if err != nil {
-		return err
-	}
-
-	_, err = tx.Exec(`CREATE UNIQUE INDEX IF NOT EXISTS roundIdx ON StateProofKeys (round);`)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec("INSERT INTO schema (tablename, version) VALUES (?, ?)", keystoreTableSchemaName, keystoreSchemaVersion)
-
-	return err
-}
-
 // GetVerifier can be used to store the commitment and verifier for this signer.
 func (s *Signer) GetVerifier() *Verifier {
 	return s.SignerRecord.GetVerifier()
@@ -281,16 +222,6 @@ func (s *SignerInRound) getMerkleTreeIndex(round uint64) uint64 {
 //	return count, err
 //}
 
-// Restore loads Signer from given database, as well as restoring PersistenKeystore (where the actual keys are stored)
-func (s *Signer) Restore(store db.Accessor) (err error) {
-	//keystore, err := RestoreKeystore(store)
-	//if err != nil {
-	//	return
-	//}
-	//s.keyStore = keystore
-	return
-}
-
 // GetKey retrieves key from memory if exists
 func (s *Signer) GetKey(round uint64) *crypto.GenericSigningKey {
 	idx := roundToIndex(s.FirstValid, round, s.Interval)
@@ -299,51 +230,6 @@ func (s *Signer) GetKey(round uint64) *crypto.GenericSigningKey {
 	}
 
 	return &s.signatureAlgorithms[idx]
-}
-
-// FetchKey returns the SigningKey and round for a specified index from the StateProof DB
-// TODO: add unit test
-func (s *Signer) FetchKey(id uint64, store db.Accessor) (*crypto.GenericSigningKey, uint64, error) {
-	var keyB []byte
-	var round uint64
-	key := &crypto.GenericSigningKey{}
-
-	err := store.Atomic(func(ctx context.Context, tx *sql.Tx) error {
-		row := tx.QueryRow("SELECT key,round FROM StateProofKeys WHERE id = ?", id)
-		err := row.Scan(&keyB, &round)
-		if err != nil {
-			return fmt.Errorf("failed to select stateProof key for round %d : %w", round, err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return nil, 0, err // fmt.Errorf("PersistentKeystore.GetKey: %w", err)
-	}
-
-	err = protocol.Decode(keyB, key)
-	if err != nil {
-		return nil, 0, err // fmt.Errorf("PersistentKeystore.GetKey: %w", err)
-	}
-
-	return key, round, nil
-}
-
-// CountKeys couts the number of rows in StateProofKeys table
-func (s *Signer) CountKeys(store db.Accessor) int {
-	var count int
-	err := store.Atomic(func(ctx context.Context, tx *sql.Tx) error {
-		row := tx.QueryRow("SELECT COUNT(*) FROM StateProofKeys")
-		err := row.Scan(&count)
-		if err != nil {
-			return fmt.Errorf("failed to count rows in table StateProofKeys : %w", err)
-		}
-		return nil
-	})
-	if err != nil {
-		return -1
-	}
-	return count
 }
 
 // RoundSecrets returns the secret keys required for the specified round as well as the public immutable data
